fix merged settings
